[ refactor ] 룰렛 결과 생성 로직 개선: useRoulette 훅 삭제 및 관련 함수 통합, URL 파라미터 추출 기능 추가, 결과 모달에 한국어 문구 추가
